# Encode and decode compacted ints